Extract configuration of attributes class from plugin_setup
To make backend specs more modular and decoupled from global plugin
setup, we'd like to have a common configuration helper, but it's
currently stuck in the middle of the plugin_setup helper.

By extracting it, we make it clearer which part of the plugin setup is
more specific to testing plugins, and which part is not.